+++ doc/ReleaseNotes.md
@@ -1,4 +1,12 @@
-﻿### 1.0.8
+﻿### 1.0.10
+
+ * Improve FSharp.Core resolution.
+
+### 1.0.9
+
+ * added FSharpAssembly.LoadFiles API.
+
+### 1.0.8
 
  * Support for custom fsi settings object ("fsi")